refactor(react): relocate old notification components to deprecated
